refactor(a2a): reuse shared provider auth schema
